+++ server/app.py
@@ -24,6 +24,8 @@ from liveblog.syndication.blogs import blogs_blueprint as syndication_blogs_blue
 from liveblog.marketplace.marketer import marketers_blueprint
 
 from liveblog.analytics.analytics import analytics_blueprint
+from liveblog.items.items import drag_and_drop_blueprint
+from liveblog.posts.posts import blog_posts_blueprint
 
 
 from superdesk.factory import get_app as superdesk_app
@@ -86,6 +88,12 @@ def get_app(config=None):
 
     # Marketplace.
     app.register_blueprint(marketers_blueprint)
+
+    # Drag and drop
+    app.register_blueprint(drag_and_drop_blueprint)
+
+    # New posts endpoint
+    app.register_blueprint(blog_posts_blueprint)
 
     return app
