+++ server/liveblog/blogs/embeds.py
@@ -14,10 +14,8 @@ import jinja2
 import json
 import logging
 import os
-import pymongo
 
 import superdesk
-from bson import ObjectId
 from bson.json_util import dumps as bson_dumps
 from eve.io.mongo import MongoJSONEncoder
 from flask import current_app as app
@@ -25,6 +23,7 @@ from flask import json, render_template, request, url_for
 from liveblog.themes import UnknownTheme
 from superdesk import get_resource_service
 from superdesk.errors import SuperdeskApiError
+from liveblog.client_modules.client_modules import Blog
 
 from .app_settings import BLOGLIST_ASSETS, BLOGSLIST_ASSETS_DIR, THEMES_ASSETS_DIR
 from .utils import is_relative_to_current_folder, get_template_file_name, get_theme_json
@@ -50,87 +49,6 @@ class ThemeTemplateLoader(jinja2.BaseLoader):
         with open(path) as f:
             source = f.read()
         return source, path, lambda: mtime == os.path.getmtime(path)
-
-
-class Blog:
-    """
-    Utility class to fetch blog data directly from mongo collections.
-    """
-    order_by = ('_updated', '_created')
-    default_order_by = '_updated'
-    sort = ('asc', 'desc')
-    default_sort = 'desc'
-
-    def __init__(self, blog):
-        if isinstance(blog, (str, ObjectId)):
-            blog = get_resource_service('client_blogs').find_one(_id=blog, req=None)
-
-        self._blog = blog
-        self._posts = get_resource_service('client_posts')
-
-    def _validate_order_by(self, order_by):
-        if order_by not in self.order_by:
-            raise ValueError(order_by)
-
-    def _validate_sort(self, sort):
-        if sort not in self.sort:
-            raise ValueError(sort)
-
-    def _posts_lookup(self, sticky=None, highlight=None):
-        filters = [
-            {'blog': {'$eq': self._blog['_id']}},
-            {'post_status': {'$eq': 'open'}},
-            {'deleted': {'$eq': False}}
-        ]
-        if sticky:
-            filters.append({'sticky': {'$eq': sticky}})
-        if highlight:
-            filters.append({'highlight': {'$eq': highlight}})
-        return {'$and': filters}
-
-    def posts(self, sticky=None, highlight=None, order_by=default_order_by, sort=default_sort, page=1, limit=25,
-              wrap=False):
-        # Validate parameters.
-        self._validate_sort(sort)
-        self._validate_order_by(order_by)
-
-        # Fetch total.
-        results = self._posts.find(self._posts_lookup(sticky, highlight))
-        total = results.count()
-
-        # Get sorting direction.
-        if sort == 'asc':
-            sort = pymongo.ASCENDING
-        else:
-            sort = pymongo.DESCENDING
-
-        # Fetch posts, do pagination and sorting.
-        skip = limit * (page - 1)
-        results = results.skip(skip).limit(limit).sort(order_by, sort)
-        posts = []
-        for doc in results:
-            if 'groups' not in doc:
-                continue
-
-            for group in doc['groups']:
-                if group['id'] == 'main':
-                    for ref in group['refs']:
-                        ref['item'] = get_resource_service('archive').find_one(req=None, _id=ref['residRef'])
-            posts.append(doc)
-
-        if wrap:
-            # Wrap in python-eve style data structure
-            return {
-                '_items': posts,
-                '_meta': {
-                    'page': page,
-                    'total': total,
-                    'max_results': limit
-                }
-            }
-        else:
-            # Return posts.
-            return posts
 
 
 def collect_theme_assets(theme, assets=None, template=None):
