+++ server/liveblog/client_modules/client_modules.py
@@ -1,3 +1,7 @@
+import pymongo
+import json
+
+from bson import ObjectId
 from liveblog.blogs.blogs import BlogsResource
 from superdesk.services import BaseService
 from liveblog.posts.posts import PostsService, PostsResource, BlogPostsService, BlogPostsResource
@@ -5,11 +9,16 @@ from superdesk.users.users import UsersResource
 from superdesk.metadata.utils import item_url
 from flask import current_app as app
 from liveblog.items.items import ItemsResource, ItemsService
-from flask import request
 from liveblog.common import check_comment_length
 from superdesk.resource import Resource
 from eve.utils import config
+from superdesk import get_resource_service
+from flask import Blueprint, request, make_response
+from flask_cors import CORS
 
+
+blog_posts_blueprint = Blueprint('blog_posts', __name__)
+CORS(blog_posts_blueprint)
 
 class ClientUsersResource(Resource):
     datasource = {
@@ -165,3 +174,103 @@ class ClientBlogPostsService(BlogPostsService):
             docs = super().get(req, lookup)
             app.blog_cache.set(blog_id, cache_key, docs)
         return docs
+
+
+class Blog:
+    """
+    Utility class to fetch blog data directly from mongo collections.
+    """
+    order_by = ('_updated', '_created')
+    default_order_by = '_updated'
+    sort = ('asc', 'desc')
+    default_sort = 'desc'
+
+    def __init__(self, blog):
+        if isinstance(blog, (str, ObjectId)):
+            blog = get_resource_service('client_blogs').find_one(_id=blog, req=None)
+
+        self._blog = blog
+        self._posts = get_resource_service('client_posts')
+
+    def _validate_order_by(self, order_by):
+        if order_by not in self.order_by:
+            raise ValueError(order_by)
+
+    def _validate_sort(self, sort):
+        if sort not in self.sort:
+            raise ValueError(sort)
+
+    def _posts_lookup(self, sticky=None, highlight=None):
+        filters = [
+            {'blog': {'$eq': self._blog['_id']}},
+            {'post_status': {'$eq': 'open'}},
+            {'deleted': {'$eq': False}}
+        ]
+        if sticky:
+            filters.append({'sticky': {'$eq': sticky}})
+        if highlight:
+            filters.append({'highlight': {'$eq': highlight}})
+        return {'$and': filters}
+
+
+    def posts(self, sticky=None, highlight=None, order_by=default_order_by, sort=default_sort, page=1, limit=25,
+              wrap=False):
+        # Validate parameters.
+        self._validate_sort(sort)
+        self._validate_order_by(order_by)
+
+        # Fetch total.
+        results = self._posts.find(self._posts_lookup(sticky, highlight))
+        total = results.count()
+
+        # Get sorting direction.
+        if sort == 'asc':
+            sort = pymongo.ASCENDING
+        else:
+            sort = pymongo.DESCENDING
+
+        # Fetch posts, do pagination and sorting.
+        skip = limit * (page - 1)
+        results = results.skip(skip).limit(limit).sort(order_by, sort)
+        posts = []
+        for doc in results:
+            if 'groups' not in doc:
+                continue
+
+            for group in doc['groups']:
+                if group['id'] == 'main':
+                    for ref in group['refs']:
+                        ref['item'] = get_resource_service('archive').find_one(req=None, _id=ref['residRef'])
+            posts.append(doc)
+
+        if wrap:
+            # Wrap in python-eve style data structure
+            return {
+                '_items': posts,
+                '_meta': {
+                    'page': page,
+                    'total': total,
+                    'max_results': limit
+                }
+            }
+        else:
+            # Return posts.
+            return posts
+
+
+@blog_posts_blueprint.route('/api/v2/client_blogs/<blog_id>/posts', methods=['GET'])
+def get_blog_posts(blog_id):
+    args = request.args
+
+    posts_service = get_resource_service('posts')
+    lookup = {'blog': ObjectId(blog_id)}
+    cursor = posts_service.get(req=None, lookup=lookup)
+
+    documents = []
+    for document in cursor:
+        documents.append(document)
+
+    return make_response(json.dumps(documents), 200)
+
+
+    # sticky highlight max_results page order_by order
